requirements.txt
huey
celery
kombu
gunicorn
Flask
Werkzeug
itsdangerous
click
Jinja2
MarkUpSafe
Wheel
google-ads
psycopg2-binary
chardet
xlrd
numpy
scipy
pandas
sklearn
glob2
redis
